Refactor Generate button to SplitButton in palette navbar
Replaces the Generate button with a SplitButton component in PaletteNavbar, providing options for generating a new palette and selecting the generation method. Updates tooltip text and adds styling props for button classes. Also updates editor-todo.md to mark related tasks as completed and add a new palette creation bug to the todo list.